jitrestaurant/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
mysql-connector-python==8.0.26
Werkzeug==2.0.1
python-dotenv==0.19.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
protobuf==4.25.3 